split sizes.rs into smaller managable files
