+++ src/sizes/mod.rs
@@ -1,0 +1,572 @@
+//! Supported array sizes: [`typenum::Unsigned`] types with an [`ArraySize`] impl.
+//!
+//! We support the following array sizes by default:
+//!
+//! - 0-512
+//! - 528-1024 (multiples of 16)
+//! - 2048, 4096, 8192
+//!
+//! When the `extra-sizes` feature is enabled: 1040-4064 (multiples of 32)
+
+/// Additional typenum size aliases beyond what are normally provided.
+///
+/// These are defined using their component bits rather than `Add` to avoid conflicting impls.
+#[cfg(feature = "extra-sizes")]
+#[allow(missing_docs)]
+mod extra_sizes;
+mod macros;
+
+#[cfg(feature = "extra-sizes")]
+pub use extra_sizes::*;
+use macros::*;
+
+impl_array_sizes_with_import! {
+    0 => U0,
+    1 => U1,
+    2 => U2,
+    3 => U3,
+    4 => U4,
+    5 => U5,
+    6 => U6,
+    7 => U7,
+    8 => U8,
+    9 => U9,
+    10 => U10,
+    11 => U11,
+    12 => U12,
+    13 => U13,
+    14 => U14,
+    15 => U15,
+    16 => U16,
+    17 => U17,
+    18 => U18,
+    19 => U19,
+    20 => U20,
+    21 => U21,
+    22 => U22,
+    23 => U23,
+    24 => U24,
+    25 => U25,
+    26 => U26,
+    27 => U27,
+    28 => U28,
+    29 => U29,
+    30 => U30,
+    31 => U31,
+    32 => U32,
+    33 => U33,
+    34 => U34,
+    35 => U35,
+    36 => U36,
+    37 => U37,
+    38 => U38,
+    39 => U39,
+    40 => U40,
+    41 => U41,
+    42 => U42,
+    43 => U43,
+    44 => U44,
+    45 => U45,
+    46 => U46,
+    47 => U47,
+    48 => U48,
+    49 => U49,
+    50 => U50,
+    51 => U51,
+    52 => U52,
+    53 => U53,
+    54 => U54,
+    55 => U55,
+    56 => U56,
+    57 => U57,
+    58 => U58,
+    59 => U59,
+    60 => U60,
+    61 => U61,
+    62 => U62,
+    63 => U63,
+    64 => U64,
+    65 => U65,
+    66 => U66,
+    67 => U67,
+    68 => U68,
+    69 => U69,
+    70 => U70,
+    71 => U71,
+    72 => U72,
+    73 => U73,
+    74 => U74,
+    75 => U75,
+    76 => U76,
+    77 => U77,
+    78 => U78,
+    79 => U79,
+    80 => U80,
+    81 => U81,
+    82 => U82,
+    83 => U83,
+    84 => U84,
+    85 => U85,
+    86 => U86,
+    87 => U87,
+    88 => U88,
+    89 => U89,
+    90 => U90,
+    91 => U91,
+    92 => U92,
+    93 => U93,
+    94 => U94,
+    95 => U95,
+    96 => U96,
+    97 => U97,
+    98 => U98,
+    99 => U99,
+    100 => U100,
+    101 => U101,
+    102 => U102,
+    103 => U103,
+    104 => U104,
+    105 => U105,
+    106 => U106,
+    107 => U107,
+    108 => U108,
+    109 => U109,
+    110 => U110,
+    111 => U111,
+    112 => U112,
+    113 => U113,
+    114 => U114,
+    115 => U115,
+    116 => U116,
+    117 => U117,
+    118 => U118,
+    119 => U119,
+    120 => U120,
+    121 => U121,
+    122 => U122,
+    123 => U123,
+    124 => U124,
+    125 => U125,
+    126 => U126,
+    127 => U127,
+    128 => U128,
+    129 => U129,
+    130 => U130,
+    131 => U131,
+    132 => U132,
+    133 => U133,
+    134 => U134,
+    135 => U135,
+    136 => U136,
+    137 => U137,
+    138 => U138,
+    139 => U139,
+    140 => U140,
+    141 => U141,
+    142 => U142,
+    143 => U143,
+    144 => U144,
+    145 => U145,
+    146 => U146,
+    147 => U147,
+    148 => U148,
+    149 => U149,
+    150 => U150,
+    151 => U151,
+    152 => U152,
+    153 => U153,
+    154 => U154,
+    155 => U155,
+    156 => U156,
+    157 => U157,
+    158 => U158,
+    159 => U159,
+    160 => U160,
+    161 => U161,
+    162 => U162,
+    163 => U163,
+    164 => U164,
+    165 => U165,
+    166 => U166,
+    167 => U167,
+    168 => U168,
+    169 => U169,
+    170 => U170,
+    171 => U171,
+    172 => U172,
+    173 => U173,
+    174 => U174,
+    175 => U175,
+    176 => U176,
+    177 => U177,
+    178 => U178,
+    179 => U179,
+    180 => U180,
+    181 => U181,
+    182 => U182,
+    183 => U183,
+    184 => U184,
+    185 => U185,
+    186 => U186,
+    187 => U187,
+    188 => U188,
+    189 => U189,
+    190 => U190,
+    191 => U191,
+    192 => U192,
+    193 => U193,
+    194 => U194,
+    195 => U195,
+    196 => U196,
+    197 => U197,
+    198 => U198,
+    199 => U199,
+    200 => U200,
+    201 => U201,
+    202 => U202,
+    203 => U203,
+    204 => U204,
+    205 => U205,
+    206 => U206,
+    207 => U207,
+    208 => U208,
+    209 => U209,
+    210 => U210,
+    211 => U211,
+    212 => U212,
+    213 => U213,
+    214 => U214,
+    215 => U215,
+    216 => U216,
+    217 => U217,
+    218 => U218,
+    219 => U219,
+    220 => U220,
+    221 => U221,
+    222 => U222,
+    223 => U223,
+    224 => U224,
+    225 => U225,
+    226 => U226,
+    227 => U227,
+    228 => U228,
+    229 => U229,
+    230 => U230,
+    231 => U231,
+    232 => U232,
+    233 => U233,
+    234 => U234,
+    235 => U235,
+    236 => U236,
+    237 => U237,
+    238 => U238,
+    239 => U239,
+    240 => U240,
+    241 => U241,
+    242 => U242,
+    243 => U243,
+    244 => U244,
+    245 => U245,
+    246 => U246,
+    247 => U247,
+    248 => U248,
+    249 => U249,
+    250 => U250,
+    251 => U251,
+    252 => U252,
+    253 => U253,
+    254 => U254,
+    255 => U255,
+    256 => U256,
+    257 => U257,
+    258 => U258,
+    259 => U259,
+    260 => U260,
+    261 => U261,
+    262 => U262,
+    263 => U263,
+    264 => U264,
+    265 => U265,
+    266 => U266,
+    267 => U267,
+    268 => U268,
+    269 => U269,
+    270 => U270,
+    271 => U271,
+    272 => U272,
+    273 => U273,
+    274 => U274,
+    275 => U275,
+    276 => U276,
+    277 => U277,
+    278 => U278,
+    279 => U279,
+    280 => U280,
+    281 => U281,
+    282 => U282,
+    283 => U283,
+    284 => U284,
+    285 => U285,
+    286 => U286,
+    287 => U287,
+    288 => U288,
+    289 => U289,
+    290 => U290,
+    291 => U291,
+    292 => U292,
+    293 => U293,
+    294 => U294,
+    295 => U295,
+    296 => U296,
+    297 => U297,
+    298 => U298,
+    299 => U299,
+    300 => U300,
+    301 => U301,
+    302 => U302,
+    303 => U303,
+    304 => U304,
+    305 => U305,
+    306 => U306,
+    307 => U307,
+    308 => U308,
+    309 => U309,
+    310 => U310,
+    311 => U311,
+    312 => U312,
+    313 => U313,
+    314 => U314,
+    315 => U315,
+    316 => U316,
+    317 => U317,
+    318 => U318,
+    319 => U319,
+    320 => U320,
+    321 => U321,
+    322 => U322,
+    323 => U323,
+    324 => U324,
+    325 => U325,
+    326 => U326,
+    327 => U327,
+    328 => U328,
+    329 => U329,
+    330 => U330,
+    331 => U331,
+    332 => U332,
+    333 => U333,
+    334 => U334,
+    335 => U335,
+    336 => U336,
+    337 => U337,
+    338 => U338,
+    339 => U339,
+    340 => U340,
+    341 => U341,
+    342 => U342,
+    343 => U343,
+    344 => U344,
+    345 => U345,
+    346 => U346,
+    347 => U347,
+    348 => U348,
+    349 => U349,
+    350 => U350,
+    351 => U351,
+    352 => U352,
+    353 => U353,
+    354 => U354,
+    355 => U355,
+    356 => U356,
+    357 => U357,
+    358 => U358,
+    359 => U359,
+    360 => U360,
+    361 => U361,
+    362 => U362,
+    363 => U363,
+    364 => U364,
+    365 => U365,
+    366 => U366,
+    367 => U367,
+    368 => U368,
+    369 => U369,
+    370 => U370,
+    371 => U371,
+    372 => U372,
+    373 => U373,
+    374 => U374,
+    375 => U375,
+    376 => U376,
+    377 => U377,
+    378 => U378,
+    379 => U379,
+    380 => U380,
+    381 => U381,
+    382 => U382,
+    383 => U383,
+    384 => U384,
+    385 => U385,
+    386 => U386,
+    387 => U387,
+    388 => U388,
+    389 => U389,
+    390 => U390,
+    391 => U391,
+    392 => U392,
+    393 => U393,
+    394 => U394,
+    395 => U395,
+    396 => U396,
+    397 => U397,
+    398 => U398,
+    399 => U399,
+    400 => U400,
+    401 => U401,
+    402 => U402,
+    403 => U403,
+    404 => U404,
+    405 => U405,
+    406 => U406,
+    407 => U407,
+    408 => U408,
+    409 => U409,
+    410 => U410,
+    411 => U411,
+    412 => U412,
+    413 => U413,
+    414 => U414,
+    415 => U415,
+    416 => U416,
+    417 => U417,
+    418 => U418,
+    419 => U419,
+    420 => U420,
+    421 => U421,
+    422 => U422,
+    423 => U423,
+    424 => U424,
+    425 => U425,
+    426 => U426,
+    427 => U427,
+    428 => U428,
+    429 => U429,
+    430 => U430,
+    431 => U431,
+    432 => U432,
+    433 => U433,
+    434 => U434,
+    435 => U435,
+    436 => U436,
+    437 => U437,
+    438 => U438,
+    439 => U439,
+    440 => U440,
+    441 => U441,
+    442 => U442,
+    443 => U443,
+    444 => U444,
+    445 => U445,
+    446 => U446,
+    447 => U447,
+    448 => U448,
+    449 => U449,
+    450 => U450,
+    451 => U451,
+    452 => U452,
+    453 => U453,
+    454 => U454,
+    455 => U455,
+    456 => U456,
+    457 => U457,
+    458 => U458,
+    459 => U459,
+    460 => U460,
+    461 => U461,
+    462 => U462,
+    463 => U463,
+    464 => U464,
+    465 => U465,
+    466 => U466,
+    467 => U467,
+    468 => U468,
+    469 => U469,
+    470 => U470,
+    471 => U471,
+    472 => U472,
+    473 => U473,
+    474 => U474,
+    475 => U475,
+    476 => U476,
+    477 => U477,
+    478 => U478,
+    479 => U479,
+    480 => U480,
+    481 => U481,
+    482 => U482,
+    483 => U483,
+    484 => U484,
+    485 => U485,
+    486 => U486,
+    487 => U487,
+    488 => U488,
+    489 => U489,
+    490 => U490,
+    491 => U491,
+    492 => U492,
+    493 => U493,
+    494 => U494,
+    495 => U495,
+    496 => U496,
+    497 => U497,
+    498 => U498,
+    499 => U499,
+    500 => U500,
+    501 => U501,
+    502 => U502,
+    503 => U503,
+    504 => U504,
+    505 => U505,
+    506 => U506,
+    507 => U507,
+    508 => U508,
+    509 => U509,
+    510 => U510,
+    511 => U511,
+    512 => U512,
+    528 => U528,
+    544 => U544,
+    560 => U560,
+    576 => U576,
+    592 => U592,
+    608 => U608,
+    624 => U624,
+    640 => U640,
+    656 => U656,
+    672 => U672,
+    688 => U688,
+    704 => U704,
+    720 => U720,
+    736 => U736,
+    752 => U752,
+    768 => U768,
+    784 => U784,
+    800 => U800,
+    816 => U816,
+    832 => U832,
+    848 => U848,
+    864 => U864,
+    880 => U880,
+    896 => U896,
+    912 => U912,
+    928 => U928,
+    944 => U944,
+    960 => U960,
+    976 => U976,
+    992 => U992,
+    1008 => U1008,
+    1024 => U1024,
+    2048 => U2048,
+    4096 => U4096,
+    8192 => U8192,
+}
